Make Member extends PartialMember, add asFullMember method
